refactor: move unit tests from src/__tests__/ to tests/unit/
Consistent test layout: tests/unit/ and tests/integration/.
